+++ tests/event/listener_test.php
@@ -1,20 +1,20 @@
 <?php
 /**
 *
-* Google Analytics extension for the phpBB Forum Software package.
+* Baidu Tongji extension for the phpBB Forum Software package.
 *
 * @copyright (c) 2014 phpBB Limited <https://www.phpbb.com>
 * @license GNU General Public License, version 2 (GPL-2.0)
 *
 */
 
-namespace phpbb\googleanalytics\tests\event;
+namespace phpbb\baidutongji\tests\event;
 
 require_once __DIR__ . '/../../../../../includes/functions_acp.php';
 
 class listener_test extends \phpbb_test_case
 {
-	/** @var \phpbb\googleanalytics\event\listener */
+	/** @var \phpbb\baidutongji\event\listener */
 	protected $listener;
 
 	/** @var \phpbb\config\config */
@@ -37,7 +37,7 @@ class listener_test extends \phpbb_test_case
 
 		// Load/Mock classes required by the event listener class
 		$this->config = new \phpbb\config\config(array(
-			'googleanalytics_id' => 'UA-000000-01',
+			'baidutongji_id' => 'UA-000000-01',
 			'ga_anonymize_ip' => 0,
 		));
 		$this->template = $this->getMockBuilder('\phpbb\template\template')
@@ -54,7 +54,7 @@ class listener_test extends \phpbb_test_case
 	*/
 	protected function set_listener()
 	{
-		$this->listener = new \phpbb\googleanalytics\event\listener(
+		$this->listener = new \phpbb\baidutongji\event\listener(
 			$this->config,
 			$this->template,
 			$this->user
@@ -79,42 +79,42 @@ class listener_test extends \phpbb_test_case
 			'core.acp_board_config_edit_add',
 			'core.page_header',
 			'core.validate_config_variable',
-		), array_keys(\phpbb\googleanalytics\event\listener::getSubscribedEvents()));
+		), array_keys(\phpbb\baidutongji\event\listener::getSubscribedEvents()));
 	}
 
 	/**
-	* Test the load_google_analytics event
+	* Test the load_baidu_tongji event
 	*/
-	public function test_load_google_analytics()
+	public function test_load_baidu_tongji()
 	{
 		$this->set_listener();
 
 		$this->template->expects($this->once())
 			->method('assign_vars')
 			->with(array(
-				'GOOGLEANALYTICS_ID'		=> $this->config['googleanalytics_id'],
-				'GOOGLEANALYTICS_TAG'		=> $this->config['googleanalytics_tag'],
-				'GOOGLEANALYTICS_USER_ID'	=> $this->user->data['user_id'],
+				'BAIDUTONGJI_ID'		=> $this->config['baidutongji_id'],
+				'BAIDUTONGJI_TAG'		=> $this->config['baidutongji_tag'],
+				'BAIDUTONGJI_USER_ID'	=> $this->user->data['user_id'],
 				'S_ANONYMIZE_IP'			=> $this->config['ga_anonymize_ip'],
 			));
 
 		$dispatcher = new \Symfony\Component\EventDispatcher\EventDispatcher();
-		$dispatcher->addListener('core.page_header', array($this->listener, 'load_google_analytics'));
+		$dispatcher->addListener('core.page_header', array($this->listener, 'load_baidu_tongji'));
 		$dispatcher->dispatch('core.page_header');
 	}
 
 	/**
-	* Data set for test_add_googleanalytics_configs
+	* Data set for test_add_baidutongji_configs
 	*
 	* @return array Array of test data
 	*/
-	public function add_googleanalytics_configs_data()
+	public function add_baidutongji_configs_data()
 	{
 		return array(
 			array( // expected config and mode
 				'settings',
 				array('vars' => array('warnings_expire_days' => array())),
-				array('warnings_expire_days', 'legend_googleanalytics', 'googleanalytics_id', 'ga_anonymize_ip', 'googleanalytics_tag'),
+				array('warnings_expire_days', 'legend_baidutongji', 'baidutongji_id', 'ga_anonymize_ip', 'baidutongji_tag'),
 			),
 			array( // unexpected mode
 				'foobar',
@@ -135,16 +135,16 @@ class listener_test extends \phpbb_test_case
 	}
 
 	/**
-	* Test the add_googleanalytics_configs event
+	* Test the add_baidutongji_configs event
 	*
-	* @dataProvider add_googleanalytics_configs_data
+	* @dataProvider add_baidutongji_configs_data
 	*/
-	public function test_add_googleanalytics_configs($mode, $display_vars, $expected_keys)
+	public function test_add_baidutongji_configs($mode, $display_vars, $expected_keys)
 	{
 		$this->set_listener();
 
 		$dispatcher = new \Symfony\Component\EventDispatcher\EventDispatcher();
-		$dispatcher->addListener('core.acp_board_config_edit_add', array($this->listener, 'add_googleanalytics_configs'));
+		$dispatcher->addListener('core.acp_board_config_edit_add', array($this->listener, 'add_baidutongji_configs'));
 
 		$event_data = array('display_vars', 'mode');
 		$event = new \phpbb\event\data(compact($event_data));
@@ -163,45 +163,45 @@ class listener_test extends \phpbb_test_case
 	}
 
 	/**
-	* Data set for test_validate_googleanalytics_id
+	* Data set for test_validate_baidutongji_id
 	*
 	* @return array Array of test data
 	*/
-	public function validate_googleanalytics_id_data()
+	public function validate_baidutongji_id_data()
 	{
 		return array(
 			array(
 				// valid code, no error
-				array('googleanalytics_id' => 'UA-0000-00'),
+				array('baidutongji_id' => 'UA-0000-00'),
 				array(),
 			),
 			array(
 				// no code, no error
-				array('googleanalytics_id' => ''),
+				array('baidutongji_id' => ''),
 				array(),
 			),
 			array(
 				// invalid code, error
-				array('googleanalytics_id' => 'UA-00-00'),
-				array('ACP_GOOGLEANALYTICS_ID_INVALID'),
+				array('baidutongji_id' => 'UA-00-00'),
+				array('ACP_BAIDUTONGJI_ID_INVALID'),
 			),
 			array(
 				// invalid code, error
-				array('googleanalytics_id' => 'UA-00000-00000'),
-				array('ACP_GOOGLEANALYTICS_ID_INVALID'),
+				array('baidutongji_id' => 'UA-00000-00000'),
+				array('ACP_BAIDUTONGJI_ID_INVALID'),
 			),
 			array(
 				// invalid code, error
-				array('googleanalytics_id' => 'AU-0000-00'),
-				array('ACP_GOOGLEANALYTICS_ID_INVALID'),
+				array('baidutongji_id' => 'AU-0000-00'),
+				array('ACP_BAIDUTONGJI_ID_INVALID'),
 			),
 			array(
 				// invalid code, error
-				array('googleanalytics_id' => 'foo-bar-foo'),
-				array('ACP_GOOGLEANALYTICS_ID_INVALID'),
+				array('baidutongji_id' => 'foo-bar-foo'),
+				array('ACP_BAIDUTONGJI_ID_INVALID'),
 			),
 			array(
-				// no googleanalytics_id, no error
+				// no baidutongji_id, no error
 				array('foo' => 'bar'),
 				array(),
 			),
@@ -209,20 +209,20 @@ class listener_test extends \phpbb_test_case
 	}
 
 	/**
-	* Test the validate_googleanalytics_id event
+	* Test the validate_baidutongji_id event
 	*
-	* @dataProvider validate_googleanalytics_id_data
+	* @dataProvider validate_baidutongji_id_data
 	*/
-	public function test_validate_googleanalytics_id($cfg_array, $expected_error)
+	public function test_validate_baidutongji_id($cfg_array, $expected_error)
 	{
 		$this->set_listener();
 
 		$config_name = key($cfg_array);
-		$config_definition = array('validate' => 'googleanalytics_id');
+		$config_definition = array('validate' => 'baidutongji_id');
 		$error = array();
 
 		$dispatcher = new \Symfony\Component\EventDispatcher\EventDispatcher();
-		$dispatcher->addListener('core.validate_config_variable', array($this->listener, 'validate_googleanalytics_id'));
+		$dispatcher->addListener('core.validate_config_variable', array($this->listener, 'validate_baidutongji_id'));
 
 		$event_data = array('cfg_array', 'config_name', 'config_definition', 'error');
 		$event = new \phpbb\event\data(compact($event_data));